+++ covid cases age.R
@@ -10,7 +10,8 @@ pacman::p_load(
   zoo,
   reshape2,
   lubridate,
-  plotly
+  plotly,
+  pals
 )
 
 # IMPORT DATASETS ----
@@ -28,13 +29,21 @@ covid_cases_age_combined <- subset(covid_cases_age_combined, age!="00_59")
 covid_cases_age_combined <- subset(covid_cases_age_combined, age!="60+")
 covid_cases_age_combined <- subset(covid_cases_age_combined, age!="unassigned")
 
-# restrict data to last two months and remove columns we don't want
+# restrict data to last months and remove columns we don't want
 covid_cases_age_combined <- subset(covid_cases_age_combined, date > today() - days(31), select = c("areaName", "date", "age", "cases"))
 
 # define the date format
 covid_cases_age_combined$date = as.Date(covid_cases_age_combined$date, "%Y-%m-%d")
 
-# PLOT DATA ----
+# create separate dataset for percentage distribution
+covid_cases_age_percents <- subset(covid_cases_age_combined, date > today() - days(10))
+
+covid_cases_age_percents <- covid_cases_age_percents %>%
+  group_by(areaName, date) %>%
+  mutate(per = 100 * cases / sum(cases)) %>%
+  ungroup
+
+# PLOT DATA LINE ----
 
 # create plot and geom
 covid_cases_age_plot <- ggplot() +
@@ -58,3 +67,26 @@ covid_cases_age_dynamic_plot
 # save to daily file
 orca(covid_cases_age_dynamic_plot, file = "dorset_age_cases.png")
 htmlwidgets::saveWidget(as_widget(covid_cases_age_dynamic_plot), "age_cases.html")
+
+# PLOT DATA BAR CHART PERCENTAGE DISTRIBUTION ----
+
+# create plot and geom
+
+covid_cases_age_stack_plot <- ggplot() +
+  geom_bar(data = covid_cases_age_percents, aes(x = date, y = per, fill = age), position = "fill", stat = "identity") +
+  #  geom_text(data = covid_cases_age_percents, aes(x = date, y = per, label = per), size = 4) +
+  xlab("Date") +
+  ylab("New cases") +
+  scale_fill_manual(name = "Age group", values = as.vector(warmcool(19))) +
+  scale_x_date(date_labels = "%d %B", date_breaks = "1 day") +
+  scale_y_continuous(labels = scales::percent_format(scale = 100)) +
+  labs(color = "Age bands") +
+  ggtitle("Dorset - new covid cases by age group distribution") +
+  labs(caption = paste("Data from UK Health Security Agency / https://coronavirus.data.gov.uk. Plotted", Sys.time(), sep = " ")) +
+  facet_grid( ~ areaName) +
+  theme_bw()
+
+covid_cases_age_stack_plot
+
+# save to daily file
+ggsave("dorset_age_cases_percentage.png", width = 16.6, height = 8.65, units = "in")